1. ec200s模块可以用shell指令进行完成网络连接 2. 但是状态机初始化的时候,不能使用malloc( )动态分配内存,否则串口将无法收到消息,目前使用静态分配内存解决 3.ec20的状态机需要完善 4.自动化连网过程的状态机需要完善 5. queue.c循环队列如何将元素类型自定义需要思考一下? 6.结构体中定义函数,如何定义使用自定义参数类型和参数数量呢?

+++ DD_Gateway_FactoryTest/HAL/Inc/queue.h
@@ -51,7 +51,7 @@ int queue_pull(Queue_HandleTypeDef *self,int *temp,uint16_t len);
 /**
  * @brief  弹出队列len个元素（要从队列中删除）
  */
-int queue_pop(Queue_HandleTypeDef *self,int *temp,uint16_t len);
+int queue_pop(Queue_HandleTypeDef *self,uint8_t *temp,uint16_t len);
 /**
  * @brief 判断队列是否满
  */
